move GetApplication() function
