+++ internal/controller/promotion/argocd.go
@@ -11,7 +11,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
-	libArgoCD "github.com/akuity/kargo/internal/argocd"
 	argocd "github.com/akuity/kargo/internal/controller/argocd/api/v1alpha1"
 	"github.com/akuity/kargo/internal/logging"
 )
@@ -213,7 +212,7 @@ func getApplicationFn(
 		namespace string,
 		name string,
 	) (*argocd.Application, error) {
-		return libArgoCD.GetApplication(ctx, argoClient, namespace, name)
+		return argocd.GetApplication(ctx, argoClient, namespace, name)
 	}
 }
