+++ internal/controller/stages/stages.go
@@ -19,9 +19,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
-	"github.com/akuity/kargo/internal/argocd"
 	"github.com/akuity/kargo/internal/controller"
-	argocdapi "github.com/akuity/kargo/internal/controller/argocd/api/v1alpha1"
+	argocd "github.com/akuity/kargo/internal/controller/argocd/api/v1alpha1"
 	"github.com/akuity/kargo/internal/kargo"
 	"github.com/akuity/kargo/internal/kubeclient"
 	"github.com/akuity/kargo/internal/logging"
@@ -61,7 +60,7 @@ type reconciler struct {
 		client client.Client,
 		namespace string,
 		name string,
-	) (*argocdapi.Application, error)
+	) (*argocd.Application, error)
 
 	// Freight qualification:
